change time_execution to double

+++ database/migrations/2022_05_24_021629_add_time_execution_to_questions.php
@@ -14,7 +14,7 @@ class AddTimeExecutionToQuestions extends Migration
     public function up()
     {
         Schema::table('questions', function (Blueprint $table) {
-            $table->float('time_execution')->nullable();
+            $table->double('time_execution')->nullable();
         });
     }
